refactor: CardInfoUiState 로직 변경
